Detect unreachable steps
This patch raises UnreachableStep when constructing a flow class that
has provably unreachable steps. Such steps are probably mistakes and
should be detected by the framework.

This doesn't handle subclasses well, in the sense that if a subclass
somehow decides to alter step connections there's no way to say "this
step is dead". Such facility can be added later on though.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>